02_chatbot_with_chainlit_streaming_tools/pyproject.toml
[project]
name = "ai-agent-with-chailnit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
]
